third_party/WebKit/LayoutTests/bluetooth/server/connect/garbage-collection-ran-during-success.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Garbage Collection ran during a connect call that ' +
   'succeeds. Should not crash.';
let device, fake_peripheral;
promise_test(() => getDiscoveredHealthThermometerDevice()
  .then(_ => ({device, fake_peripheral} = _))
  .then(() => fake_peripheral.setNextGATTConnectionResponse({
    code: HCI_SUCCESS,
  }))
  .then(() => {
    device.gatt.connect();
    device = undefined;
    fake_peripheral = undefined;
    return runGarbageCollection();
  })
  .then(() => navigator.bluetooth.test.allResponsesConsumed())
  .then(consumed => assert_true(consumed)), test_desc);
</script>
